Move the container classes to the util package
