+++ pygradle-plugin/src/main/groovy/com/linkedin/gradle/python/util/ZipappContainer.java
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.linkedin.gradle.python.tasks;
-
+package com.linkedin.gradle.python.util;
 
 public interface ZipappContainer extends ApplicationContainer {
     /**
